// src/app/dto/user.dto.ts
export class UserDto {
    username?: string;
    password?: string;
    roles?: number[];
    firstname?: string;
    lastname?: string;
    gender?: string;
    phone?: string;
    provinceId?: number;
    districtId?: number;
    village?: string;
    verifyCode?: string;
    userId?: number;
    branchId?: number;
    restaurantId?: number;
}